-Added MagnetizePIDPoseAlign -Replaced PIDPoseAlign with MagnetizePIDPoseAlign in AutoScoreCommands
MagnetizePIDPoseAlign takes in the same parameters as PIDPoseAlign except it needs controller for magnetization inputs. MagnetizationPIDPoseAlign calculates distance like PIDPoseAlign and subtracts controller inputs so the driver has power over the magnetization. The power of magnetization can be changed by modifying the magnetizationPower in the MagnetizePIDPoseAlign file (I found 5.0 to work fine, but we should change it based on our driver's preferences).